Merge in SMAB/qa_automation from RegFix_CA to pre-release

* commit '21a6580813e0f4f5328c6ef73c42d0e45ffda4b2':
  Regression Fixes
  Regression Fix
  Added code coverage for SMAB-15286

+++ src/test/java/com/apas/Tests/BPPManagement/BPP_GeneralAircraftStatementProcessing_Test.java
@@ -62,8 +62,7 @@ public class BPP_GeneralAircraftStatementProcessing_Test extends TestBase {
 		softAssert = new SoftAssertion();
 		rollYear = "2022";
 		objWorkItemHomePage = new WorkItemHomePage(driver);
-		objEfileHomePage.updateRollYearStatus("Open", rollYear);
-
+		
 	}
 		
 	/**
@@ -75,11 +74,11 @@ public class BPP_GeneralAircraftStatementProcessing_Test extends TestBase {
 	public void BPP_GeneralAircraft_AircraftAnnualSettings(String loginUser) throws Exception {
 		
 		//Delete BPP Annual Settings
-		objSalesforceAPI.deleteBPPAnnualSettings(rollYear);
+		objSalesforceAPI.deleteCurrentRollYearBPPAnnualSettings(rollYear);
 		 
 		//Delete the existing WI from system before importing files
-        String query = "select id from Work_Item__c where Request_Type__c = 'BPP - Annual Setting - BPP-Annual Setting'";
-        objSalesforceAPI.delete("Work_Item__c", query);
+        String query = "select id from Work_Item__c where Request_Type__c = 'BPP - Annual Setting - BPP-Annual Setting' and (Status__c = 'In Pool' Or Status__c = 'In Progress')";
+        if (query != null) objSalesforceAPI.delete("Work_Item__c", query);
         
         //Generate Reminder WI
         objSalesforceAPI.generateReminderWorkItems(SalesforceAPI.REMINDER_WI_CODE_BPP_AIRCRAFT_ANNUAL_SETTINGS);
@@ -119,7 +118,7 @@ public class BPP_GeneralAircraftStatementProcessing_Test extends TestBase {
 				"SMAB-T3840: Validate Land PI default % value on the Annual Settings record");
         
         //Get the WI id and navigate to it
-        String workItemQuery = "Select id, Name from Work_Item__c where Request_Type__c = 'BPP - Annual Setting - BPP-Annual Setting' limit 1";
+        String workItemQuery = "Select id, Name from Work_Item__c where Name = '" + aircraftAnnualSettingsWorkItem  + "' limit 1";
 		String workItemId = objSalesforceAPI.select(workItemQuery).get("Id").get(0);
         
 		String executionEnv = System.getProperty("region");
@@ -191,5 +190,148 @@ public class BPP_GeneralAircraftStatementProcessing_Test extends TestBase {
         //Log out from the application
         objBppManagementPage.logout();
 	}
+	
+	
+	/**
+	 * DESCRIPTION: Verify creation of Airport Listing Work Item 
+	 * and Import Airport Listing data
+	 */
+	
+	@Test(description = "SMAB-T4449,SMAB-T4266,SMAB-T4267: Verify creation of Airport Listing Work Item and Import Airport Listing data", dataProvider = "loginPrincipalUser", dataProviderClass = DataProviders.class, groups = {"Regression", "GeneralAircraft", "BPPManagement"})
+	public void BPP_GeneralAircraft_AirportListingImport(String loginUser) throws Exception {
+		
+		//Login to the APAS application 
+		objEfileHomePage.login(loginUser);
+		
+		//Delete or Revert the existing data from system before importing files
+		String query = "Select id From E_File_Import_Log__c where File_type__c = 'Airport Listing' and Import_Period__C='" + rollYear + "' and File_Source__C = 'Airports' and (Status__c = 'Imported' Or Status__c = 'Approved')";
+		objSalesforceAPI.update("E_File_Import_Log__c", query, "Status__c", "Reverted");
+		objSalesforceAPI.deleteAircraftListingWIs();
 
+		//Generate Reminder WI
+        objSalesforceAPI.generateReminderWorkItems(SalesforceAPI.REMINDER_WI_CODE_BPP_EFILE);
+
+		//Fetch the excel file to be imported for Airport Listing
+		String fileName = System.getProperty("user.dir") + testdata.BPP_AIRPORT_LISTING_DATA;		
+				
+		//Open the Work Item Home Page
+        objBppManagementPage.searchModule(modules.HOME);
+        objBppManagementPage.waitForElementToBeClickable(20, objWorkItemHomePage.lnkTABInPool);
+		objWorkItemHomePage.Click(objWorkItemHomePage.lnkTABInPool);
+		objWorkItemHomePage.waitForElementToBeVisible(20,objWorkItemHomePage.acceptWorkItemButton);
+       
+        //Verify that only one WI is generated
+        String airportListingRequestType = "BPP - Airport Listing Import";
+        int airportListingWorkItemCount = objWorkItemHomePage.getWorkItemCount(airportListingRequestType, objWorkItemHomePage.TAB_IN_POOL);
+        softAssert.assertEquals(airportListingWorkItemCount, 1, 
+        		"SMAB-T4266: Validate that only one WI for Airport Listing is generated");
+        
+        //Verify the WorkPool name on the WI
+        String aircraftAnnualSettingsWorkItem = objWorkItemHomePage.getWorkItemName(airportListingRequestType,objWorkItemHomePage.TAB_IN_POOL);
+        softAssert.assertEquals(objWorkItemHomePage.getGridDataForRowString(airportListingRequestType).get("Work Pool").get(0), "Aircraft Work Pool", 
+        		"SMAB-T4266: Validate the Work Pool name on WI for Airport Listing");
+
+        //Accepting the work item and opening the link under 'Action' Column
+        ReportLogger.INFO("Accept the WI");
+        objWorkItemHomePage.acceptWorkItem(aircraftAnnualSettingsWorkItem);
+        objWorkItemHomePage.Click(objWorkItemHomePage.inProgressTab);
+        objWorkItemHomePage.openActionLink(aircraftAnnualSettingsWorkItem);
+        Thread.sleep(3000); //Allow upload button to appear
+        objPage.verifyElementExists("//span[contains(.,'Upload Files')]");
+        
+        //Get the WI id and navigate to it
+        String workItemQuery = "Select id from Work_Item__c where Name = '" + aircraftAnnualSettingsWorkItem  + "' limit 1";
+		String workItemId = objSalesforceAPI.select(workItemQuery).get("Id").get(0);
+        
+		String executionEnv = System.getProperty("region");
+		driver.navigate().to("https://smcacre--"+executionEnv.toLowerCase()+
+				 ".lightning.force.com/lightning/r/Work_Item__c/"+workItemId+"/view");
+		
+		objBppManagementPage.waitForElementToBeClickable(20, objWorkItemHomePage.detailsTab);
+		objWorkItemHomePage.Click(objWorkItemHomePage.detailsTab);
+		objBppManagementPage.waitForElementToBeClickable(20, objWorkItemHomePage.inProgressOptionInTimeline);
+		
+		softAssert.assertEquals(objWorkItemHomePage.getFieldValueFromAPAS(objWorkItemHomePage.wiStatus), "In Progress",
+				"SMAB-T4266: Validate the status of WI for Airport Listing");
+		
+		//Validate the WI details
+		ReportLogger.INFO("Validate the WI details");
+		objWorkItemHomePage.Click(objWorkItemHomePage.detailsTab);
+		objWorkItemHomePage.waitForElementToBeVisible(objWorkItemHomePage.referenceDetailsLabel, 20);
+		
+		softAssert.assertContains(objBppManagementPage.getFieldValueFromAPAS(objBppManagementPage.relatedActionLabel),"Airport Listing Import",
+				"SMAB-T4267: Validate the Related Action label value on WI for Airport Listing");
+        softAssert.assertContains(objBppManagementPage.getFieldValueFromAPAS(objBppManagementPage.workPoolLabel),"Aircraft Work Pool",
+				"SMAB-T4267: Validate the Work Pool name on WI for Airport Listing");
+        softAssert.assertContains(objBppManagementPage.getFieldValueFromAPAS(objBppManagementPage.actionLabel),"Airport Listing Import",
+				"SMAB-T4267: Validate the Action label value on WI for Airport Listing");
+        
+        //Click on Related action link
+		objWorkItemHomePage.Click(objWorkItemHomePage.reviewLink);
+		String parentWindow = driver.getWindowHandle();
+		objWorkItemHomePage.switchToNewWindow(parentWindow);
+		
+		//Uploading the Airport Listing file having error and success records
+		objEfileHomePage.uploadFile(fileName + "AirportListingExcelSheet.xlsx");		
+		
+		//Waiting for the status of the file to be converted to Imported
+		ReportLogger.INFO("Waiting for Status of the imported file to be converted to Imported");
+		objPage.waitForElementTextToBe(objEfileHomePage.statusImportedFile, "Imported", 360);
+		
+		//Opening the Efile Import Logs module
+		objEfileHomePage.searchModule(modules.EFILE_IMPORT_LOGS);
+		HashMap<String, ArrayList<String>> importLogsGridData = objEfileHomePage.getGridDataInHashMap(1, 1);
+		
+		//Import Logs grid validation for the imported Airport Listing file
+		softAssert.assertEquals(importLogsGridData.get("Name").get(0),"Airport Listing :Airports :" + rollYear,
+				"SMAB-T4449: Validate the name of imported file on import logs page");
+		softAssert.assertEquals(importLogsGridData.get("Status").get(0),"Imported", 
+				"SMAB-T4449: Validate if status of imported file is imported on import logs page");
+		
+		//Opening the Efile Import Transactions module
+		objEfileHomePage.searchModule(modules.EFILE_IMPORT_TRANSACTIONS);
+		HashMap<String, ArrayList<String>> importTransactionsGridData = objEfileHomePage.getGridDataInHashMap(1, 1);
+		
+		//Import Transactions grid validation for the imported Airport Listing file
+		softAssert.assertEquals(importTransactionsGridData.get("E-File Import Log").get(0),"Airport Listing :Airports :" + rollYear,
+				"SMAB-T4449: Validate the name of imported file on import transactions page");
+		softAssert.assertEquals(importTransactionsGridData.get("Status").get(0),"Imported", 
+				"SMAB-T4449: Validate if status of imported file is imported on import transactions page");
+		
+		//Opening the e-file intake module
+		objEfileHomePage.searchModule(modules.EFILE_INTAKE);
+		
+		//Selecting the File Type and Source and clicking on 'View All' link 
+		objEfileHomePage.selectFileAndSource("Airport Listing", "Airports");
+		objPage.Click(objEfileHomePage.viewLink);
+		objPage.waitForElementToBeVisible(objEfileHomePage.errorRowSection,30);
+		
+		//Validating that correct number of records are present in error row section after file import
+		String expectedNoOfErrorRecords = "2";    //Imported File has 2 error records in each table			
+		
+		//Validation of number of records in error row section
+		String actualNoOfErrorRecords = objEfileHomePage.getCountOfRowsFromErrorOrImportedRowsSection("ERROR");
+		softAssert.assertEquals(actualNoOfErrorRecords, expectedNoOfErrorRecords, 
+				"SMAB-T4449: Validate if correct number of records are displayed in Error Row Section after file import");
+				
+		//Validation of number of records in imported row section					
+		String expectedImportedRowsCount = "15";	
+		String actualImportedRowsCount = objEfileHomePage.getCountOfRowsFromErrorOrImportedRowsSection("IMPORTED");
+		softAssert.assertEquals(actualImportedRowsCount, expectedImportedRowsCount, 
+				"SMAB-T4449: Validation if correct number of records are displayed in Imported Row Section after file import");
+			
+		//********* Validate Discard Functionality *********	
+		//Validation for Records discard functionality from Review and Approve Page
+		ReportLogger.INFO("Validate that error records can be discarded from Review and Approve Data Page");
+		objEfileHomePage.discardErrorRecords("1");
+			
+		//Validating the number of records in the error row section after discarding a record
+		int updatedCount = Integer.parseInt(expectedNoOfErrorRecords) - 1;
+		actualNoOfErrorRecords = objEfileHomePage.getCountOfRowsFromErrorOrImportedRowsSection("ERROR");
+		String updatedRecordsInErrorRowPostDelete = Integer.toString(updatedCount);		
+		softAssert.assertEquals(actualNoOfErrorRecords, updatedRecordsInErrorRowPostDelete, 
+				"SMAB-T4449: Validation if correct number of records are displayed in Error Row Section after discarding a record");
+
+		objBppManagementPage.logout();
+	}
 }
